requirements.txt
pathway[all]
python-dotenv==1.0.1
mpmath==1.3.0
litellm>=1.35
google-generativeai>=0.4.0
sentence-transformers>=2.5.1
fastapi[all]
uvicorn[standard]
pdfminer.six==20221105
groq